atom-editor/atom-editor-git: Various improvements
Changing the source (like patching and seding) should be done in the prepare() function (if possible).
In this case it's possible and the changes don't get lost

Additionally copy the manually created .desktop over the one that the atom build process generates (non working) to avoid two .desktop files in the finished package.

Remove the conflicts line from pointed release PKGBuild. The git PKGBuild is already conflicting the other one. This already is sufficient to handle the update process between git and non-git package (or in that case the conflict). And unnecessary lines should be removed from a PKGBuild.

The pkgver() result is changed to reflect the prefered way.